+++ includes/requirements.php
@@ -19,9 +19,9 @@
 require_once TW_DIR_LIB . 'aihrus-framework/aihrus-framework.php';
 
 
-function tw_requirements_check() {
+function tw_requirements_check( $force_check = false ) {
 	$check_okay = get_transient( 'tw_requirements_check' );
-	if ( $check_okay !== false ) {
+	if ( empty( $force_check ) && $check_okay !== false ) {
 		return $check_okay;
 	}